+++ kitty/options/definition.py
@@ -2892,6 +2892,20 @@ map('Scroll to bottom',
     only="macos",
     )
 
+map('Scroll to previous shell prompt',
+    'scroll_to_previous_prompt kitty_mod+x scroll_to_prompt -1',
+    long_text='''
+Requires :ref:`shell_integration` to work.
+'''
+    )
+
+map('Scroll to next shell prompt',
+    'scroll_to_next_prompt kitty_mod+d scroll_to_prompt 1',
+    long_text='''
+Requires :ref:`shell_integration` to work.
+'''
+    )
+
 map('Browse scrollback buffer in less',
     'show_scrollback kitty_mod+h show_scrollback',
     long_text='''